fix(stage-ui): share Steam-aware linked-account linking with stage-web/tamagotchi
Widening OAuthProvider to include 'steam' broke account-settings-page.vue's
handleLinkProvider typing (shared by stage-web and stage-tamagotchi), which
was missed when the same fix was applied to ui-server-auth's profile page.

Promotes the Steam /link-social workaround into stage-ui's
use-linked-accounts module as withSteamLinking(), so both consumers share
one implementation instead of each hand-rolling a client wrapper.
