+++ apps/ui-server-auth/src/pages/profile.vue
@@ -2,7 +2,7 @@
 import type { ProfileUser } from '../modules/profile'
 
 import { defaultSignInProviders } from '@proj-airi/stage-ui/components/auth'
-import { useLinkedAccounts } from '@proj-airi/stage-ui/composables'
+import { useLinkedAccounts, withSteamLinking } from '@proj-airi/stage-ui/composables'
 import { SERVER_URL } from '@proj-airi/stage-ui/libs/server'
 import { Avatar, Button, FieldInput } from '@proj-airi/ui'
 import { computed, onMounted, reactive, shallowRef } from 'vue'
@@ -17,7 +17,7 @@ import {
   trackPasswordResetRequested,
   trackSignedOut,
 } from '../modules/analytics'
-import { getAuthClient, toLinkedAccountsClient } from '../modules/auth-client'
+import { getAuthClient } from '../modules/auth-client'
 import { requestPasswordReset } from '../modules/email-password'
 import {
   changePassword,
@@ -95,7 +95,7 @@ const {
   unlink: unlinkLinkedProvider,
   link: linkLinkedProvider,
 } = useLinkedAccounts({
-  client: toLinkedAccountsClient(getAuthClient({ apiServerUrl })),
+  client: withSteamLinking(getAuthClient({ apiServerUrl })),
   isAuthenticated,
   describeError: describeProfileError,
   messages: {
